chore: include more options for sveltekit chore: remove manifestTransform from sveltekit example
